+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Make hotkey optional for plugin commands and toolbar](https://github.com/siyuan-note/siyuan/pull/16676)
 * [Improve IProtyleOptions](https://github.com/siyuan-note/plugin-sample/issues/38)
+* [Improve error handling and security for plugin data storage methods](https://github.com/siyuan-note/siyuan/pull/16717)
 
 ## v1.1.7 2025-12-23